Chrome bug? Fix rotation when resizing
If you change the size of a rotated control, it would not be centered
correctly. Changing the rotation would snap it back to its correct
position. I'm guessing it's at least a Chrome bug with using SVG
transform and CSS transform-origin. Just use SVG transform with the
origin specified in the rotate command.